feat: requests re-configured

+++ LSP.Entity/Concrete/ClassroomCapacity.cs
@@ -1,11 +1,14 @@
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using LSP.Core;
 using LSP.Entity.Abstract.Common;
+using Microsoft.EntityFrameworkCore;
 
 namespace LSP.Entity.Concrete
 {
     public class ClassroomCapacity : BaseEntity, IEntity
     {
-        public byte Id { get; set; }
+        public short Id { get; set; }
         public short Capacity { get; set; }
     }
 }
