+++ LSP.Entity/Concrete/ClassroomType.cs
@@ -1,3 +1,5 @@
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using LSP.Core;
 using LSP.Entity.Abstract.Common;
 
@@ -5,7 +7,7 @@ namespace LSP.Entity.Concrete
 {
     public class ClassroomType : BaseEntity, IEntity
     {
-        public byte Id { get; set; }
+        public short Id { get; set; }
         public string Name { get; set; } = "UNKNOWN";
     }
 }
